u-boot: fix boot index propagation
During an update, in case of failure of active partition after
several attempts (bootcount max count triggered), the system
needs to reboot on previous_active partition, so u-boot needs
to read FWU_INFO backup register (fwu metadata partition is
not updated by tf-a).
It was the case for stm32mp1 series, but stm32mp2 series was
reading boot index through metadata partition, which was not
correct.

Change-Id: Ifd2fef32a504b1d712b357dc529b0e9c0e3c70d1